requirements.txt
langchain
langchain-core
langchain-aws
langchain-community
streamlit
python-dotenv
boto3